+++ src/run/constants.py
@@ -156,6 +156,6 @@ SEVEN_ZIP_PATH = read_ini_config(ini_file_path, "unzip_config", "SEVEN_ZIP_PATH"
 # unzip zip file password
 PASSWORD = read_ini_config(ini_file_path, "unzip_config", "PASSWORD")
 
-sis_server_version = "v1.1.2-beta.1.240401"
-build_date = "2024-04-01 18:00"
-publish_date = "2024-04-02 18:30"
+sis_server_version = "v1.1.2-beta.2.240410"
+build_date = "2024-04-10 18:00"
+publish_date = "2024-04-11 18:30"